artinya ap bg messi

+++ labs/lab04/exercise2/lab-4-2.py
@@ -3,7 +3,7 @@ if income <= 50000.0:
     totalTax = 0
 else:
     if totalTax <= 100000.0:
-        totalTax = float(50000 * 1/ 100) + ((income - 100000) * 2 / 100)
+        totalTax = (50000 * 1/ 100) + ((income - 100000) * 2 / 100)
     else:
         totalTax = ((income - 50000) * 1 / 100)
 print(totalTax)
